add: search comment controller, user and post model functions

+++ app/Http/Controllers/CommentController.php
@@ -29,6 +29,23 @@ class CommentController extends Controller
         return response()->json(['message' => 'Comment added successfully'], 201);
     }
 
+//edit comment
+    public function editComment(Request $request, $commentId)
+    {
+
+        $comment = Comment::find($commentId);
+        $this->authorize('edit', $comment);
+
+        if (!$comment) {
+            return response()->json(['message' => 'Comment not found'], 404);
+        }
+
+        $comment->comment_content = $request->input('content'); 
+        $comment->save();
+
+        return response()->json(['message' => 'Comment updated successfully'], 200);
+    }
+
 // Get post comments
     public function getComments($postId)
     {
@@ -51,8 +68,21 @@ class CommentController extends Controller
         if (!$comment) {
             return response()->json(['message' => 'Comment not found'], 404);
         }
+        $this->authorize('delete', $comment);	
 
         $comment->delete();
         return response()->json(['message' => 'Comment deleted successfully'], 200);
+    }
+
+    public function searchComments(Request $request)
+    {
+        $input = $request->get('search') ? $request->get('search').':*' : "*";
+
+        $visiblePosts = Post::where('is_public', true)->pluck('id')->toArray();
+        $comments = Comment::whereRaw("tsvectors @@ to_tsquery(?)", [$input])
+                            ->orderByRaw("ts_rank(tsvectors, to_tsquery(?)) ASC", [$input])
+                            ->whereIn('post_id', $visiblePosts)->get();
+
+        return response()->json($comments, 200);
     }
 }
